Remove FootballTermsDetector and create FootballFilter.
This FootballFilter class is much more intimate with the filtering
process. It has access to tags provided by a NER, allowing it to make
decisions based on the types of tags and keywords identified.

As a result, the AnalyzedText class is now more stripped out. It acts as
a container for the original text and the identified named entities and
tags.